requirements.txt
# Core Backend Dependencies
langchain-groq>=0.0.1
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
chromadb>=0.4.24
reportlab>=4.0.0
pillow>=10.0.0
psutil>=5.9.0
pandas>=2.0.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# LinkedIn Scraper Dependencies
python-jobspy>=1.1.82
selenium>=4.15.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
pyperclip>=1.8.2
requests>=2.31.0
beautifulsoup4>=4.12.2
numpy>=1.26.3
pydantic>=2.3.0
tls-client>=1.0.1
markdownify==0.13.1
regex>=2024.4.28 